+++ CHANGELOG.md
@@ -94,6 +94,8 @@ https://github.com/shadow/shadow/issues/2718
 * Stopped overriding libc's `getcwd` with an incorrect wrapper that was
 returning `-1` instead of `NULL` on errors.
 
+* A call to `epoll_ctl` with an unknown operation will return `EINVAL`.
+
 Raw changes since v2.5.0:
 
 * [Merged PRs](https://github.com/shadow/shadow/pulls?q=is%3Apr+merged%3A%3E2023-03-23T18%3A20-0400)